requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
gunicorn
psycopg2-binary